Set up new serialization base
